[#1] refactor: redux store 분리 및 redux-persist 세팅
