+++ package.json
@@ -23,6 +23,7 @@
     "react-redux": "^8.0.7",
     "react-router-dom": "^6.12.0",
     "react-scripts": "5.0.1",
+    "redux-persist": "^6.0.0",
     "styled-components": "^5.3.10",
     "swr": "^2.1.5",
     "typescript": "^4.9.5",
